Refactoring of directories for packaging
