+++ pyecog2/annotation_table_widget.py
