+++ pyecog2/classifier.py
