+++ pyecog2/coding_tests/ParameterTreeTest.py
