+++ pyecog2/coding_tests/VideoPlayer.py
