+++ pyecog2/coding_tests/WPTacquisition.py
