+++ pyecog2/coding_tests/dock_tests.py
