+++ pyecog2/coding_tests/pyqtgraph_examples.py
