+++ pyecog2/coding_tests/save_settings.py
