+++ pyecog2/coding_tests/testing_menu.py
