+++ pyecog2/deprecated/main.QtCore import *
