+++ pyecog2/deprecated/old__forref_PyecogCurvePlotItem.py
