+++ pyecog2/descriptions_etc.txt
