+++ pyecog2/feature_extractor.py
