+++ pyecog2/file_selection_dialog_for_notebooks.py
