+++ pyecog2/hmm_pyecog.py
