+++ pyecog2/icons/folder.png
