+++ pyecog2/icons/folder_paper.png
