+++ pyecog2/icons/laboratory-mouse.png
