+++ pyecog2/icons/research.png
