+++ pyecog2/icons/wave.png
