+++ pyecog2/live_plot.py
