+++ pyecog2/paired_graphics_view.py
