+++ pyecog2/spectrotools.py
