+++ pyecog2/tree_widget.py
